move tweet info
